refactor(dashboard): extract getResourceHealth to lib/resource-health.ts
Moves the pure function out of QASession.tsx to fix the
react-refresh/only-export-components lint warning and make it
directly importable in tests.
